Code/RasaWorkshop/zomatobot/cli_version/requirements.txt
## Setup py requirements
jsonpickle==1.0.0
six==1.11.0
redis==2.10.6
fakeredis==0.10.3
future==0.17.1
numpy~=1.16.0
typing==3.6.6
ruamel.yaml==0.15.85
requests~=2.2
tensorflow==1.15.0
h5py==2.7.1
apscheduler==3.5.1
tqdm==4.29.1
ConfigArgParse==0.13.0
networkx==2.2
pykwalify==1.7.0
coloredlogs==10.0
flask==1.0.2
flask_cors==3.0.7
scikit-learn==0.20.2
rasa_nlu~=0.14.4
colorhash==1.0.2
pika==0.12.0
jsonschema==2.6.0
packaging~=18.0
gevent==1.4
pytz==2018.9
rasa-core-sdk~=0.12.2
rasa_core~=0.13.2
pymongo==3.7.2
python-dateutil==2.7.5
spacy~=2.0.18
sklearn_crfsuite==0.3.6
keras-applications==1.0.6
keras-preprocessing==1.0.5
